[#5333 4/6] Region snapshot replacement GC
This commit adds a "region snapshot replacement garbage collect"
background task that will scan for region snapshot replacement requests
in the ReplacementDone state and trigger a saga to delete the old
snapshot volume populated in the region snapshot replacement start saga.

The "garbage collect" saga added here does nothing except change the
state of the region snapshot replacement request and invoke a volume
delete sub-saga. At the end of this saga the record's state is changed
to Running.