[gateway] turn API into a trait (#6145)
Somewhat large, but with no functional changes -- this is pure code movement.

In a few spots I had to move away from `From` impls, so that gateway-types does
not depend on gateway-sp-comms. I think those changes are for the best.

As a bonus, sp-sim no longer has to depend on omicron-gateway, cutting down the
number of build units for it from 600 to 468. That's a really nice improvement.