[xtask] move downloader to a separate binary (#6171)
While iterating on xtask I noticed that it could be pretty slow to
compile. Turned out that a lot of the dependencies were actually coming from the
download tool.

Move the downloader into a separate binary.

On Linux, this results in the number of build units for xtask dropping
from 254 to 122. That's a pretty nice improvement.