Reconfigurator: Teach planner about SNAT IPs
Prior to this change, the planner expected all blueprints to have
fully-exclusive external IP addresses. This isn't compatible with #6037,
where RSS now hands out SNAT IPs with distinct port ranges but the same
IP address.

A big chunk of this work is necessary to support boundary NTP zone
planning, but that isn't included in this PR, so those bits are marked
with `#[cfg(test)]`.

Fixes #6194.