[Oximeter] Start Cluster based integration tests based on Clickward (#6131)
Add a test that dynamically reconfigures clickhouse replicas and keeper
clusters while writing data and querying it via OxQL. I've run this test
locally in a loop for over an hour. It takes ~30-50 seconds to run, but
that appears to be the nature of the beast. Reducing the client timeout
makes it flaky, as well as the total timeout that we wait for replicas
and keepers to become ready.

This PR also ensures that all replicated measurement and field tables
have corresponding `Distributed` tables, along with the `Schema` table.
It simplifies the creation of `Distributed` tables using the `AS`
keyword and therefore resolves
https://github.com/oxidecomputer/omicron/issues/4089. Additionally, the
replicated sql files are split into two, where each file contains
disjoint definitions. We only load the minimal set in `db-init-1.sql`
for the test in this PR, and load both for production and the existing
tests.