Wait for "running snapshots" to be running!
The snapshot create saga doesn't actually wait for running snapshots to
be created, it simply requests them! This has been mostly fine up until
now, but Nexus should actually wait for the state of the running
snapshot to transition from Requested to Created, meaning that the
Crucible agent has taken all the steps to actually start the appropriate
read-only downstairs.