Centralise IP address object address retrieval and existence checks. (#6204)
Following [illumos 16677](https://www.illumos.org/issues/16677), the
error message that indicates that
an address object does not exist has been changed to be
consistent, whereas it was previously one of two different messages
or sometimes even completely blank.

A couple of places in omicron relied on the specific error
message so this change centralises the check, and updates it
to cater for illumos pre- and post-16677. Longer term this
could be replaced by bindings to libipadm.