Reconfigurator: Move external IP allocation/deallocation into a transaction conditional on the current blueprint (#6240)
See #6239 and
https://github.com/oxidecomputer/omicron/pull/6207#discussion_r1701982773
for additional context, but briefly summarizing:

* Prior to this PR, blueprint execution would deallocate and allocate
external IP and NIC records independently
* If two Nexus instances are trying to realize two different blueprints,
this could result in them stomping on each other

Andrew suggested that we put the external IP resource set under a
generation number, similar to how we mark generations for zones and DNS
config. However, in both of those cases, there's an external entity
(sled-agent or the DNS zone, respectively) that's ultimately responsible
for rejecting updates with an old generation number. In this external
networking case, there is no other entity - other parts of Nexus consume
this information directly from CRDB. In chat, we discussed that we think
it's equivalent to putting the set of allocation/deallocation operations
inside a transaction that's conditional on the blueprint being executing
still being the current target.

Most of this PR is moving the `external_networking` module from
`nexus-reconfigurator-execution` into `nexus-db-queries` where it can be
executed inside the new
`blueprint_ensure_external_networking_resources()` method. This method
opens a transaction, performs a `SELECT ... FOR UPDATE` to confirm that
the blueprint we're ensuring is the current target, and then does the
entire set of deallocation / allocation operations in that transaction.
I added a test with some kinda gnarly `#[cfg(test)]` bits to control
execution flow to show how the `SELECT ... FOR UPDATE` works: if the
blueprint isn't the target, we immediately fail; if we are the target,
any attempt to _change_ the target will be queued by CRDB until we've
finished our transaction.

Fixes #6239