internal NTP: resolve boundary NTP sources from DNS in addition to being told explicitly (#6050)
This PR adds a special internal DNS name
`boundary.ntp.control-plane.oxide.internal` which resolves to a set of
AAAA records, one for each boundary NTP zone. We pass this name to
chrony via a `pool` directive in its config file, allowing it to find
the boundary NTP servers via internal DNS.

This PR does not remove the explicit boundary NTP server names from
either the sled-agent -> zone-setup or the zone-setup -> chrony config
paths. Assuming this PR ships as part of R10, we can come back and
remove those after R10 is out the door. We can't do both in one release,
because we need to establish the new DNS name (via setting a new
blueprint) in R10, at which point we can remove the explicit server
names in R11, because NTP has to sync before we get the opportunity to
modify DNS.

Fixes #4791.