[sled-agent] remove NexusClientWithResolver (#6283)
Previously, we would carry around a DNS resolver -- however, the resolver has
been unused for a while. And in fact there's a warning about being careful
around using it.

Because the type was public, the rustc dead code detector didn't figure that
out until my change to switch to an API trait in #6159 exposed this.

Remove `NexusClientWithResolver`, adding a couple of free functions to make a
`NexusClient` instead.