Add a new `ClickhouseServer` Omicron Zone
The new zone type reflects the zone with which we'll deploy clickhouse
server nodes in a replicated setup. We decided on using a new zone type
rather than adding a boolean field to the existing `Clickhouse` zone
type that is used for single server deployments in last Tuesday's (Aug
6, 2024) update huddle.

This is the fist part of the work to be done for replicated clickhouse
deployments that are automated via reconfigurator. As such, the actual
zone deployment is left as a `todo`.