[oximeter] Add some units from physical reality (#6296)
In service of future changes to record data from power, temperature, and
fan speed sensors in Oximeter, this branch adds some physical quantities
to the `Units` enum: `Volts`, `Amps`, `DegreesCelcius`, and `Rpm`. I've
added all of these as whole numbers of the measured quantities, with the
expectation that they will probably be recorded as floating-point. We
could consider instead using a smaller unit like `MilliAmps`, and
recording them as integers, but that introduces a bunch of dimensional
analysis that I'm not sure if we want to be doing.