[internal-dns] remove lookup_ipv6 in favor of lookup_socket_v6
`lookup_ipv6` is both buggy and easy to misuse:
* it sends an AAAA query for a domain which should have a SRV record
  - this works only because https://github.com/oxidecomputer/omicron/issues/4051
    means the SRV record is incorrectly returned, along with the desired
    AAAA in Additionals
* it looks up an IPv6 address from a SRV record *but ignores the port*.
  in places `lookup_ipv6` was used, it was paired consistently with the
  hardcoded port NEXUS_INTERNAL_PORT and matched what should be in the
  resolved SRV record. if we for example wanted to move Nexus' port (or
  start a test Nexus on an atypical port), the authoritative port number
  in the SRV response would be ignored for the hardcoded port.

lets just use the port that we told DNS we're at!

we may still want a bare IPv6 address for a service if we're testing
network reachability, for example, but we're not doing that today. and
if we need a service's IPv6 address to use with an alternate port to
access a different API, we *probably* should have a second SRV
record for that API to use instead?

(`lookup_ipv6` would be simple enough to fix and retain, but after
looking at its uses it seems it might be more trouble than it's worth
right now)