[sp-sim] Update README instructions
The instructions for running MGS currently given in `sp-sim/README.adoc`
seem to have bitrotted: the MGS binary is no longer named `gateway`, and
is now called `mgs`, and it now requires `--address <IPv6 addr>` and
`--id <UUID>` arguments as well.

Furthermore, the config file provided with the SP simulator is now
called `config.toml`, rather than `sidecar.toml`; as it contains
configurations for both a simulated Sidecar *and* two simulated Gimlets.

This commit updates the instructions in the README to match the current
reality.