[sled-agent] turn API into a trait (#6159)
Building on the changes in #6138, this PR turns the sled-agent API into a trait. This is
high-impact because it currently is part of a circular dependency with the
Nexus internal API -- this PR gets us to the point where the OpenAPI documents
can be generated independently of each other.

There is some code movement but there are no functional changes in this PR -- I
didn't use `replace` here to keep the PR small. I'll try out `replace` for some
of the types with `From` impls in a subsequent PR.

(I did change the internal `VpcFirewallRule` to `ResolvedVpcFirewallRule` to
match the other types in the destination file.)

I've also added some documentation about when `nexus-sled-agent-shared` should
be used.

Depends on #6283.