[internal-dns] remove lookup_ipv6 in favor of lookup_socket_v6 (#6320)
`lookup_ipv6` is both buggy and easy to misuse:
* it sends an AAAA query for a domain which should have a SRV record
- this works only because
https://github.com/oxidecomputer/omicron/issues/4051 means the SRV
record is incorrectly returned, along with the actually-desired AAAA for
the SRV's target in Additionals
* it looks up an IPv6 address from a SRV record *but ignores the port*.
in places `lookup_ipv6` was used, it was paired consistently with the
hardcoded port NEXUS_INTERNAL_PORT and matched what should be in the
resolved SRV record. if we for example wanted to move Nexus' port (or
start a test Nexus on an atypical port), the authoritative port number
in the SRV response would be ignored for the hardcoded port.

lets just use the port that we told DNS we're at!

we may still want a bare IPv6 address for a service if we're going to
test network reachability, for example, but we're not doing that with
this function today. this all is distinct from helpers like
`lookup_all_ipv6`.

if we need a service's IPv6 address to use with an alternate port to
access a different API, we *probably* should have a distinct SRV record
for that lookup to use instead? i've found three instances of this:
* wicket assumes the techport proxy is on the same IP as Nexus' API, but that isn't necessarily true
* we assume the CRDB admin service listens on the same IP as CRDB itself, but that doesn't have to be true
* we look up addresses for MGS via `ServiceName::Dendrite`, but there's a
`ServiceName::ManagementGatewayService`, so either that's a typo or can be made to have its own SRV records

there are some uses of `lookup_all_ipv6` that make a lot of sense still,
where we're discovering the rack's network and _really_ do not care
about the port that Dendrite happens to be on.