don't make the user specify sensor IDs in config
While working on #6354, it turned out that having to manually assign
these was really easy to mess up when you want to simulate a lot of
sensors, and there wasn't an easy way to get "real" ID values from an
inventory of a real life SP, so the config writer has to assign them
sequentially anyway. #6354 ended up using `component_details`
rather than `read_sensor`, so being able to assign sensor IDs identical
to those from Hubris turned out to not be that important, since (IIUC)
the only way to get the `SensorId`s is to look at the Hubris archive,
while *component* IDs can be discovered dynamically from the inventory
APIs.