back out "Update Rust crate tokio to 1.39.2 (#6249)" (#6356)
Tokio 1.39 updated its mio dependency to 1.0, which changed the waker impl on illumos from a self-pipe to eventfd. That has caused several issues already:

* https://github.com/oxidecomputer/helios/issues/169
* https://github.com/oxidecomputer/helios/pull/171

Based on these and the potential for other lurking issues, we're making a policy decision to roll back to 1.38 (mio 0.8) for r10. We can't be off of the train forever so we're aiming to land the 1.39 update early in the r11 cycle.

This backs out commit d7d4beaf0dbfa82c6ae0da10a6ce43b3c5a89142.