rename most of the schema fields
Initially, I tried to reuse the same names from the `gateway_sp_comms`
RPC interface for the metrics fields, since I thought it was nice to
keep the naming consistent throughout the whole stack. However, some of
these names are kind of confusing or unclear to the reader. Based on
@bnaecker's [comments][1] and some things I thought seemed kind of
unclear, I've given the metric schema fields names that I hope are more
readily understandable. In particular:

- `component` is now `hardware_component`, since `component` is way too
  generic
- `device` is now `component_kind` --- in the `gateway_sp_comms` API, a
  "component" is the name of the individual component, while a "device"
  is the _type_ of component, e.g. a TMP117 sensor. I found this kind of
  unclear.
- Changed the metrics' `name` field to `sensor`.
- Made the use of `kind`/`type` consistent by using `kind` everywhere,
  instead of having `chassis_type` and `sensor_kind`.
- Added a `chassis_` prefix to `serial`, `model`, and `revision`, to
  make it clear that those fields describe the sled/switch/power shelf,
  rather than the individual component on that chassis.