Use `BlueprintZoneConfig` in RSS service plan
In anticipation of adding more `BlueprintZoneConfig` variants with more
auxiliary information, we stop converting from `OmicronZoneConfig` to
`BlueprintZoneConfig` which is not going to be feasible for much longer.

Instead we change the one production code place we do this, RSS, to directly
construct `BlueprintZoneConfig` structs rather than do the conversion.

This has some ripple effects, and results in a new persistent v4 sled
service plan.

There is one test that still does this conversion, but the function
that does it is now moved into that test module and commented heavily.
We hope to remove it shortly.