discard samples if SP state changes mid-poll
Thanks to @jgallagher for pointing out that a SP's firmware may be
updated while we are in the middle of scraping its sensors (see [this
comment][1]). Now, we'll re-check that the SP's state hasn't changed
since the start of the poll, and throw away the current batch of samples
and retry if it has. This avoids reporting samples to Oximeter that may
have incorrect fields.

[1]: https://github.com/oxidecomputer/omicron/pull/6354#issuecomment-2302953239