SP poller tasks should never need to be restarted
Per [this conversation][1] with @jgallagher, it turns out that the
fairly complex code for restarting poller tasks is basically entirely
necessary. The `ManagementSwitch::sp` function actually only returns an
error if discovery hasn't run yet, or if the SP identifier is invalid.
Because we wait to start poller tasks until discovery has run, and we
use SP identifiers returned by the management switch, our call to that
function should never fail, so we can make the poller tasks never return
an error and not have to worry about restarting them. This makes the
code a lot simpler.

[1]: https://github.com/oxidecomputer/omicron/pull/6354#discussion_r1725727261