[sled-agent] avoid using hardcoded port number to reach MGS
this is one (and the easiest) of three cases i noted in #6363 where we
find a service by looking up a SRV record, ignoring the port in that
record, keeping IPs from that SRV lookup, and then finally pairing those
addresses with a service's hardcoded port.

in this case, we were looking up the SRV record for Dendrite, then using
the MGS port instead. this this technically works, because the two
*will* be running together, but we can just get the right port from the
get-go by looking for MGS's SRV record instead. so, do that.

this file also mentions MGS_PORT in `init_switch_config`, at which point
it's more necessary: in that case, we're configuring *our* switch zone.
so we *know* the address we want to reach - it's our own sled - and we
just need to add the port of the service we want.