Idempotently delete freed crucible regions
In the volume delete saga, separate detecting which freed crucible
regions could maybe be cleaned up from the actual deletion step. The
`find_deleted_volume_regions` function that returns which regions may be
ready for clean up joins on the volume table, and part of the deletion
step is to delete volume records - this affects the results of the
query, so store those results in the output of a previous saga node.