sled agent: index running VMMs by VMM ID, not instance ID (#6429)
Change sled agent's instance lookup tables so that Propolis jobs are
indexed by Propolis/VMM IDs instead of instance IDs.
This is a prerequisite to revisiting how the Failed instance state
works. See RFD 486 section 6.1 for all the details of why this is
needed, but very broadly: when an instance's VMM is Destroyed, we'd like
sled agent to tell Nexus that *before* the agent deregisters the
instance from the sled, for reasons described in the RFD; but if we do
that with no other changes, there's a race where Nexus may try to
restart the instance on the same sled before sled agent can update its
instance table, causing instance start to fail.

To achieve this:

- In sled agent, change the `InstanceManagerRunner`'s instance map to a
`BTreeMap<PropolisUuid, Instance>`, then clean up all the compilation
errors.
- In Nexus:
- Make callers of instance APIs furnish a Propolis ID instead of an
instance ID. This is generally very straightforward because they already
had to get a VMM record to figure out what sled to talk to.
- Change `cpapi_instances_put` to take a Propolis ID instead of an
instance ID. Regular sled agent still has both IDs, but with these
changes, simulated sled agents only have a Propolis ID to work with, and
plumbing an instance ID down to them requires significant code changes.
- Update test code:
- Unify the Nexus helper routines that let integration tests get sled
agent clients or sled IDs; now they get a single struct containing both
of those and the instance's Propolis IDs.
- Update users of the simulated agent's `poke` endpoints to use Propolis
IDs.
- Delete the "detach disks on instance stop" bits of simulated sled
agent. These don't appear to be load-bearing, they don't correspond to
any behavior in the actual sled agent (which doesn't manage disk
attachment or detachment), and it was a pain to rework them to work with
Propolis IDs.

Tests: cargo nextest.

Related: #4226 and #4872, among others.