[meta] update tokio to 1.39.3 (again) (#6452)
Tokio 1.39/mio 1.0 switches out the illumos impl to being eventfd based. For release 10 we decided that that was too risky, so we switched back to Tokio 1.38.

Now that the r10 branch has been cut, we can go back and update Tokio to 1.39.3. We'd like to land this early in the cycle to get as much soak time as possible.

See:

* #6356
* #6249
* https://github.com/oxidecomputer/helios/issues/169
* https://github.com/oxidecomputer/helios/pull/171
* #6391